requirements_true.txt
# Most come pre-packaged via the core libraries

# Refer to the base installation guide beforehand at
# https://github.com/JMoonr/LATR/blob/main/docs/install.md

# pytorch==1.8.0 torchvision==0.9.0 torchaudio==0.8.0 cudatoolkit=10.1 -c pytorch

# All packages in conda environment

# Packages from conda-forge
_openmp_mutex==4.5  # conda-forge
blas==2.125  # conda-forge
blas-devel==3.9.0  # conda-forge
cudatoolkit==10.1.243  # conda-forge
colorama==0.4.6  # conda-forge
freetype==2.12.1  # conda-forge
fvcore==0.1.5.post20221221  # conda-forge
intel-openmp==2024.2.1  # conda-forge
iopath==0.1.10  # conda-forge
jpeg==8d  # conda-forge
lerc==4.0.0  # conda-forge
libblas==3.9.0  # conda-forge
libcblas==3.9.0  # conda-forge
libdeflate==1.14  # conda-forge
libgcc==14.2.0  # conda-forge
libgfortran==14.2.0  # conda-forge
libgfortran5==14.2.0  # conda-forge
libgomp==14.2.0  # conda-forge
liblapack==3.9.0  # conda-forge
liblapacke==3.9.0  # conda-forge
libpng==1.6.43  # conda-forge
libwebp==1.2.4  # conda-forge
libwebp-base==1.2.4  # conda-forge
libwinpthread==12.0.0  # conda-forge
libxcb==1.13  # conda-forge
libzlib==1.2.13  # conda-forge
mkl==2024.2.2  # conda-forge
mkl-devel==2024.2.2  # conda-forge
mkl-include==2024.2.2  # conda-forge
ninja==1.12.1  # conda-forge
numpy==1.21.6  # conda-forge
python==3.7.0  # conda-forge
python_abi==3.7  # conda-forge

# Packages from PyPI
absl-py==2.1.0
addict==2.4.0
aiofiles==22.1.0
aiosqlite==0.19.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
attrs==24.2.0
babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
black==23.3.0
bleach==6.0.0
cached-property==1.5.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.15.1
charset-normalizer==3.4.0
click==8.1.7
comm==0.1.4
crcmod==1.7
cryptography==44.0.0
cycler==0.11.0
dataclasses==0.8
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
descartes==1.1.0
entrypoints==0.4
exceptiongroup==1.2.2
fastjsonschema==2.21.0
fire==0.7.0
flake8==3.9.2
fonttools==4.38.0
fqdn==1.5.1
geffnet==1.0.2
google-auth==2.36.0
google-auth-oauthlib==0.4.6
grpcio==1.62.3
idna==3.10
imageio==2.31.2
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.4
jmespath==0.10.0
joblib==1.3.2
json5==0.9.16
jsonpointer==3.0.0
jsonschema==4.17.3
jupyter==1.1.1
jupyter-client==7.4.9
jupyter-console==6.6.3
jupyter-core==4.12.0
jupyter-events==0.6.3
jupyter-server==1.24.0
jupyter-server-fileid==0.9.3
jupyter-server-ydoc==0.8.0
jupyter-ydoc==0.2.5
jupyterlab==3.6.8
jupyterlab-pygments==0.2.2
jupyterlab-server==2.24.0
jupyterlab-widgets==3.0.13
kiwisolver==1.4.5
llvmlite==0.36.0
lyft-dataset-sdk==0.0.8
markdown==3.4.4
markdown-it-py==2.2.0
markupsafe==2.1.5
matplotlib==3.5.2
matplotlib-inline==0.1.6
mccabe==0.6.1
mdurl==0.1.2
mistune==3.0.2
mmcv-full==1.5.0
mmdet==2.24.0
mmdet3d==1.0.0rc3
mmengine==0.9.1
mmsegmentation==0.20.0
model-index==0.1.11
munkres==1.1.4
mypy-extensions==1.0.0
nbclassic==1.1.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
nest-asyncio==1.6.0
networkx==2.2
notebook==6.5.7
notebook-shim==0.2.4
numba==0.53.0
nuscenes-devkit==1.1.10
oauthlib==3.2.2
opencv-python==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.10
ordered-set==4.1.0
ortools==9.2.9972
oss2==2.17.0
packaging==24.0
pandas==1.1.5
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.11.2
pickleshare==0.7.5
pillow==9.3.0
pip==24.0
pkgutil-resolve-name==1.3.10
platformdirs==4.0.0
plotly==5.18.0
pluggy==1.2.0
plyfile==0.9
portalocker==2.7.0
prettytable==3.7.0
prometheus-client==0.17.1
prompt-toolkit==3.0.48
protobuf==3.20.3
psutil==6.1.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycodestyle==2.7.0
pycparser==2.21
pycryptodome==3.21.0
pyflakes==2.3.1
pygments==2.17.2
pyparsing==3.1.4
pypiwin32==223
pyquaternion==0.9.9
pyrsistent==0.19.3
pytest==7.4.4
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytorch==1.8.0
pytz==2023.3
pywavelets==1.3.0
pywin32==308
pywinpty==2.0.10
pyyaml==6.0.1
pyzmq==26.2.0
pyzmq==26.2.0 # pypi
regex==2024.4.4 # pypi
requests==2.31.0 # pypi
requests-oauthlib==1.3.1 # pypi
rich==13.6.0 # pypi
rsa==4.9 # pypi
sacred==0.8.4 # pypi
scikit-image==0.19.3 # pypi
scikit-learn==0.24.2 # pypi
scipy==1.7.1 # pypi
seaborn==0.11.2 # pypi
setuptools==67.0.0 # pypi
shapely==1.8.5 # pypi
six==1.16.0 # pypi
smmap==5.0.0 # pypi
sniffio==1.3.0 # pypi
soupsieve==2.6 # pypi
sqlite==3.41.2 # conda-forge
statsmodels==0.12.2 # pypi
tabulate==0.9.0 # pypi
tensorboard==2.13.0 # pypi
tensorboard-data-server==0.6.1 # pypi
tensorboard-plugin-wit==1.8.1 # pypi
tensorflow==2.9.1 # pypi
tensorflow-estimator==2.9.0 # pypi
termcolor==2.4.0 # pypi
terminado==0.17.1 # pypi
testpath==0.6.0 # pypi
texttable==1.6.7 # pypi
threadpoolctl==2.2.0 # pypi
tifffile==2021.3.17 # pypi
torchvision==0.9.0 # pypi
tornado==6.3.3 # pypi
tqdm==4.65.0 # pypi
traitlets==5.11.2 # pypi
typer==1.5.0 # pypi
typing_extensions==4.7.1 # pypi
urllib3==2.1.0 # pypi
uvicorn==0.24.0 # pypi
wcwidth==0.2.6 # pypi
websocket-client==1.7.3 # pypi
websockets==11.0.3 # pypi
werkzeug==2.3.7 # pypi
widgetsnbextension==4.2.1 # pypi
xarray==2023.9.0 # pypi
xgboost==1.6.2 # pypi
xmltodict==1.0.1 # pypi
zipp==3.16.2 # pypi